build: finish rename inbound IHE GW functions
Ref. metriport/metriport-internal#1350
