+++ packages/lambdas/src/inbound-patient-discovery.ts
@@ -1,4 +1,4 @@
-import { processInboundPatientDiscovery } from "@metriport/core/external/carequality/pd/process-incoming-pd";
+import { processInboundPatientDiscovery } from "@metriport/core/external/carequality/pd/process-inbound-pd";
 import { MPIMetriportAPI } from "@metriport/core/mpi/patient-mpi-metriport-api";
 import { getEnvVarOrFail } from "@metriport/core/util/env-var";
 import { inboundPatientDiscoveryReqSchema } from "@metriport/ihe-gateway-sdk";
